#2 ukm  :: refaktoryzacja pakietów
Wydzielenie osobnych pakietów dla klas modelu związanych z wydarzeniami, z użytkownikiem i klas ogólnych
