+++ src/main/java/honestit/projects/events/model/common/BaseEntity.java
@@ -1,4 +1,4 @@
-package honestit.projects.events.model;
+package honestit.projects.events.model.common;
 
 import lombok.Getter;
 import lombok.Setter;
