+++ src/main/java/honestit/projects/events/model/common/Identifiable.java
@@ -1,4 +1,4 @@
-package honestit.projects.events.model;
+package honestit.projects.events.model.common;
 
 import java.io.Serializable;
