+++ src/main/java/honestit/projects/events/model/event/EventCancelReason.java
@@ -1,4 +1,4 @@
-package honestit.projects.events.model;
+package honestit.projects.events.model.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
