+++ src/main/java/honestit/projects/events/model/event/EventReview.java
@@ -1,18 +1,20 @@
-package honestit.projects.events.model;
+package honestit.projects.events.model.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import java.io.Serializable;
 
 @Embeddable
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
-public class UserDetails implements Serializable {
+public class EventReview implements Serializable {
 
-    private String firstName;
-    private String lastName;
-
+    @Column(nullable = false)
+    private Integer rating;
+    @Column(length = 0xffff)
+    private String review;
 }
